usage-limits/app/pyproject.toml
[project]
name = "usage-limits-app"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.81",
    "databricks-sql-connector>=3.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "cachetools>=5.3",
    "apscheduler>=3.10",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-mock>=3.14",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "setup", "routers", "schemas"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Pure unit tests with no external dependencies",
    "integration: Tests that exercise multiple modules together (still mocked at boundaries)",
]
addopts = "-v --tb=short --strict-markers"
pythonpath = ["."]
